apps/ai-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
httpx==0.28.1
google-generativeai==0.8.5
redis==5.3.0
pydantic==2.11.1
